order_worker/requirements.txt
pymysql==1.1.1
pika==1.3.2
psycopg2-binary==2.9.9
